+++ app/lib/l10n/gen/ouds_flutter_app_localizations_en.dart
@@ -107,9 +107,6 @@ class AppLocalizationsEn extends AppLocalizations {
   String get app_tokens_color_content_label => 'Content';
 
   @override
-  String get app_tokens_color_decorative_label => 'Decorative';
-
-  @override
   String get app_tokens_color_overlay_label => 'Overlay';
 
   @override